CorrelationTrackers.jl

CorrelationTrackers.jl package has means to do fast updates of correlation functions calculated by CorrelationFunctions.Directional module in CorrelationFunctions.jl. Correlation functions are recalculated when you change an element of the underlying array. Currently, only Directional.l2, Directional.s2 and Directional.surfsurf functions are supported.

An update after one element of the underlying system was changed is about 20000 times faster compared to a full recalculation of the correlation functions for that system.

Examples

This is a step-by-step guide on how to use CorrelationTrackers.jl:

For the beginning, create your system you want to calculate correlation functions for. Let it be two-dimensional two-phase system with size 20x30.

Pick combinations of correlation function + phase which you want to track, e.g $S_2^1(r)$, $L_2^1(r)$ and $L_2^0(r)$. A combination of function and phase is represented by AbstractTracker structures, e.g. S2Tracker(1) will track $S_2^{(1)}(r)$ function.

Create CorrelationTracker structure and access it as an ordinary array.

Obtain correlation functions at any time using functions from CorrelationFunctions.Directional. module.

using StatsBase
using PrettyTables
using Random
using CorrelationFunctions
using CorrelationTrackers

# Create our system
system = rand(MersenneTwister(348), 0:1, (20, 30))

# Suppose we want to track $S_2^1(r)$ and $L_2^0(r)$.
tracking = [S2Tracker(1), L2Tracker(1)]

# Create the tracker. It may require some time if your system is big.
tracker = CorrelationTracker(system; tracking = tracking)

# Do some phase swapping in both the original system and the tracker
system[12, 25]  = 1 - system[12, 25]
system[11, 20]  = 1 - system[11, 20]
tracker[12, 25] = 1 - tracker[12, 25]
tracker[11, 20] = 1 - tracker[11, 20]

pretty_table(stdout, hcat(Directional.s2(system,  1) |> mean,
                          Directional.l2(system,  1) |> mean,
                          Directional.s2(tracker, 1) |> mean,
                          Directional.l2(tracker, 1) |> mean);
                     header = ["S2 recalculated", "L2 recalculated",
                               "S2 tracked",      "L2 tracked"])

┌─────────────────┬─────────────────┬────────────┬────────────┐
│ S2 recalculated │ L2 recalculated │ S2 tracked │ L2 tracked │
├─────────────────┼─────────────────┼────────────┼────────────┤
│        0.491667 │        0.491667 │   0.491667 │   0.491667 │
│        0.253913 │        0.253913 │   0.253913 │   0.253913 │
│        0.236364 │        0.127273 │   0.236364 │   0.127273 │
│        0.250476 │       0.0647619 │   0.250476 │  0.0647619 │
│           0.243 │           0.036 │      0.243 │      0.036 │
│        0.238947 │       0.0189474 │   0.238947 │  0.0189474 │
│        0.236667 │       0.0111111 │   0.236667 │  0.0111111 │
│        0.247059 │      0.00705882 │   0.247059 │ 0.00705882 │
│           0.225 │         0.00375 │      0.225 │    0.00375 │
│           0.224 │      0.00266667 │      0.224 │ 0.00266667 │
└─────────────────┴─────────────────┴────────────┴────────────┘

API

CorrelationTrackers.AbstractTracker — Type

Generic type for correlation function tracker.

See also: L2Tracker, S2Tracker, SSTracker.

CorrelationTrackers.S2Tracker — Type

S2Tracker(phase)

Descriptor for two-point correlation function for the phase phase.

See also: L2Tracker, SSTracker, AbstractTracker.

CorrelationTrackers.L2Tracker — Type

L2Tracker(phase)

Descriptor for line-segment correlation function for the phase phase.

See also: S2Tracker, SSTracker, AbstractTracker.

CorrelationTrackers.SSTracker — Type

SSTracker(phase)

Descriptor for surface-surface correlation function for the phase phase.

See also: S2Tracker, L2Tracker, AbstractTracker.

CorrelationTrackers.default_trackers — Function

default_trackers(T)

Construct a vector of correlation functions which are tracked by default (that is $S_2^1(x)$, $L_2^1(x)$ and $L_2^0(x)$). T is the type of x.

CorrelationTrackers.CorrelationTracker — Type

CorrelationTracker(system   :: AbstractArray{T, N}; 
                   tracking = default_trackers(T),
                   periodic = false[, directions][, kwargs...])

Create correlation functions tracker.

Create correlation tracker for the array system. tracking is a vector of AbstractTracker structures which specify correlation functions you wish to track. periodic and direction have the same meaning as in the most functions in CorrelationFunctions.jl package. Additional arguments such as len may be passed in kwargs.

Returned tracker supports interface of AbstractArray (e.g. you can perform element-wise read and write operations).

Examples

julia> let
       system = rand(MersenneTwister(35), 0:1, (30, 10))
       tracker = CorrelationTracker(system)
       end
30×10 CorrelationTracker{Int64, 2, Matrix{Int64}}:
 0  1  0  1  1  0  0  1  1  0
 1  1  1  0  0  0  0  0  1  1
 0  0  0  0  0  0  1  1  0  1
 1  1  1  0  1  1  1  0  1  0
 0  1  0  0  1  0  0  1  1  1
 0  0  0  0  0  0  1  0  1  1
 0  0  1  0  1  1  0  1  0  1
 1  0  0  1  0  0  1  0  1  0
 0  1  1  0  0  1  1  1  1  1
 0  0  1  1  1  1  0  0  0  0
 0  0  1  1  0  0  1  1  1  0
 0  1  0  0  0  1  0  0  1  0
 1  0  0  1  0  0  1  1  0  1
 0  1  0  1  0  0  1  1  1  0
 1  1  0  1  1  1  0  1  0  1
 1  1  1  0  0  0  0  1  0  1
 1  0  0  1  0  0  1  1  1  0
 0  0  0  1  0  0  0  1  1  0
 1  0  1  0  1  0  0  0  1  0
 1  0  0  1  0  0  0  0  0  1
 1  1  1  0  1  0  1  0  1  1
 0  1  0  1  1  0  0  1  0  1
 0  0  0  1  0  0  1  1  1  1
 0  0  1  1  1  1  0  1  1  0
 1  0  1  1  0  0  0  0  0  1
 1  1  0  1  0  1  1  0  1  0
 0  1  1  0  0  1  1  0  1  0
 0  1  0  0  1  0  0  1  0  0
 1  1  0  0  1  1  0  0  0  1
 0  0  1  1  0  1  1  1  1  0

CorrelationTrackers.tracked_data — Function

tracked_data(x :: CorrelationTracker)

Return an iterator over correlation function descriptors which are tracked by the tracker.

CorrelationTrackers.tracked_length — Function

tracked_length(x :: CorrelationTracker)

Return maximal tracked correlation length

Examples

julia> tracked_length(CorrelationTracker{Int,2}(rand(0:1, (50, 100))))
25

CorrelationTrackers.tracked_directions — Function

tracked_directions(x :: CorrelationTracker)

Return directions along which correlation functions are tracked.

Examples

julia> tracked_directions(CorrelationTracker{Int,2}(rand(0:1, (50, 100))))
2-element Vector{Symbol}:
 :x
 :y

CorrelationTrackers.update_corrfns! — Function

update_corrfns!(tracker, value, index)

This function is equivalent to writing tracker[index] = value with exception that it also returns a rollback handle which can fastly bring the tracker to the previous state by calling rollback!.

See also: rollback!.

CorrelationTrackers.rollback! — Function

rollback!(tracker, token)

Bring the system to the state before update_corrfns! was called. token must be an object returned by update_corrfns!.

See also: update_corrfns!.

Caveats

Currently Directional.l2 and Directional.s2 return internal structures of supplied tracker. Do not modify them. This behavior can be changed to a safer one in the future.